+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v2.25.5
 
 * slight optimization for Heap.push_many
+* bugfix for Heap.push and a deprecation
 
 # v2.25.4